StImageViewer - added options to flip cubemap Z
with 6x1 turned ON by default and 3x2 turned OFF by default.
No option within video player - flip is disabled.

+++ StMoviePlayer/StVideo/StVideoQueue.cpp
@@ -440,7 +440,7 @@ bool StVideoQueue::init(AVFormatContext*   theFormatCtx,
             const double aPitch = (double )aSpherical->pitch / (1 << 16);
             const double aRoll  = (double )aSpherical->roll  / (1 << 16);
             myRotateDeg = (int )-aRoll;
-            theNewParams->setRotateZero((float )-aYaw, (float )aPitch, myRotateDeg);
+            theNewParams->setRotateZero((float )-aYaw, (float )aPitch, (float )myRotateDeg);
         }
     }
 #endif
